@@ -1,23 +1,22 @@
 import logo from './logo.svg';
 import './App.css';
+import Header from './components/Header';
+import { BrowserRouter as Router, Route, Routes} from 'react-router-dom';
+
+import Products from './components/Products';
+import Home from './components/Home';
 
 function App() {
   return (
     <div className="App">
-      <header className="App-header">
-        <img src={logo} className="App-logo" alt="logo" />
-        <p>
-          Edit <code>src/App.js</code> and save to reload.
-        </p>
-        <a
-          className="App-link"
-          href="https://reactjs.org"
-          target="_blank"
-          rel="noopener noreferrer"
-        >
-          Learn React
-        </a>
-      </header>
+<Router>
+    <Header/>
+    <Routes>
+    <Route path="/" exact element={<Home/>}/>
+   <Route path="/products" exact element={<Products/>} />
+
+    </Routes>
+  </Router>
     </div>
   );
 }
